Object Orientation
Converted from a procedural to object oriented style of programming.
All functionality remains the same. The functions for saving and
loading are difficult to parse and a nicer solution would be
preferable.

To Be Added
Finish typing all functions.
Implement uuid for each user and eventually public/private key
pairs for wallets.
Signed-off-by: Aaron Zolotor <azolotor@tuta.io>

 On branch master
 Changes committed:
	deleted:    PyCoin.py
	new file:   block.py
	new file:   blockchain.py
	new file:   blockchain.txt
	new file:   node.py
	new file:   transaction.py
	modified:   utility/hash_util.py
	new file:   verification.py