Wallet Implementation
Implemented wallets use public/private (RSA) keys whose public key
serves as the node id. Wallets can be loaded/saved so different node
ids are possible on a single machine. i.e. if different people were using
it to host/mine at different times. SHA512 hashes are used in the
implementation of the wallets to be consistent with the hashes that
appear elsewhere in the program.

To Be Added
Web Interface (implemented with flask)

Signed-off-by: Aaron Zolotor <azolotor@tuta.io>

 On branch master
 Changes committed:
	modified:   block.py
	modified:   blockchain.py
	modified:   blockchain.txt
	modified:   node.py
	modified:   verification.py
	new file:   wallet.py